requirements.txt
# Hivemind MCP Dependencies

# MCP SDK
mcp>=0.9.0

# API Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Web App
flask>=3.0.0

# Anthropic Claude
anthropic>=0.8.0

# Google Cloud
google-cloud-firestore>=2.13.1

# HTTP Client
httpx>=0.25.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
